Point to correct shell on Linux
The /bin/sh on Mac is a linked to /bin/bash, but on Linux it is not the same.
A few commands and syntax don't exist on /bin/sh and it failed to execute there.

This commit changes the she-bang to point to bash instead, allowing the commands to execute on Linux
and Windows WSL.

Partially closes #9