Fix key mismatch by using url
This commit fixes the backend to allow installing dat from domains.
We previously were using the canonical key of the dat repository, instead of the given url passed, and this would mismatch when looking into the folders.

The folder would be created using the given URL passed on the view, but when creating the global_apps we would store it under their canonical dat key.

This would make the app load forever, trying to open a file into a folder that doesn't exist.

This commit changes the storage to use the given url for the applications, and uses the same folder name as the key to start.
Now the application works for domains